move mjpeg to same actix-web instance
