+++ src/lib/mjpeg_streaming/mod.rs
@@ -1,4 +1,3 @@
-mod mjpeg_streaming;
 mod broadcaster;
 
-pub use self::{mjpeg_streaming::*, broadcaster::*};
+pub use self::{broadcaster::*};
