+++ src/lib/rest_api/mjpeg.html
