+++ src/lib/rest_api/rest_api.rs
@@ -5,22 +5,40 @@ use actix_cors::Cors;
 use crate::settings::AppSettings;
 use crate::lib::rest_api::services;
 use crate::lib::data_storage::DataStorage;
+use crate::lib::mjpeg_streaming::Broadcaster;
+use std::sync::{
+    Mutex,
+    mpsc::{
+        Receiver
+    }
+};
+use opencv::{
+    core::Vector,
+};
 
 pub struct Storage {
     pub data_storage: Arc<RwLock<DataStorage>>,
     pub app_settings: AppSettings,
-    pub settings_filename: String
+    pub settings_filename: String,
+    pub mjpeg_broadcaster: web::Data<Mutex<Broadcaster>>
 }
 
 #[actix_web::main]
-pub async fn start_rest_api(server_host: String, server_port: i32, data_storage: Arc<RwLock<DataStorage>>, app_settings: AppSettings, settings_filename: &str) -> std::io::Result<()> {
+pub async fn start_rest_api(server_host: String, server_port: i32, data_storage: Arc<RwLock<DataStorage>>, enable_mjpeg: bool, rx_frames_data: Receiver<Vector<u8>>, app_settings: AppSettings, settings_filename: &str) -> std::io::Result<()> {
     let bind_address = format!("{}:{}", server_host, server_port);
     println!("REST API is starting on host:port {}:{}", server_host, server_port);
     let storage = Storage{
         data_storage: data_storage,
         app_settings: app_settings,
-        settings_filename: settings_filename.to_string()
+        settings_filename: settings_filename.to_string(),
+        mjpeg_broadcaster: web::Data::new(Mutex::new(Broadcaster::default())),
     };
+
+    /* Enable MJPEG streaming server if needed */
+    if enable_mjpeg == true {
+        Broadcaster::spawn_reciever(storage.mjpeg_broadcaster.clone(), rx_frames_data);
+    }
+
     let data = web::Data::new(storage);
     HttpServer::new(move || {
         let cors = Cors::default()
@@ -33,10 +51,11 @@ pub async fn start_rest_api(server_host: String, server_port: i32, data_storage:
         App::new()
             .wrap(cors)
             .app_data(data.clone())
-            .configure(services::init_routes)
+            .configure(services::init_routes(enable_mjpeg))
     })
     .bind(&bind_address)
     .unwrap_or_else(|_| panic!("Could not bind server to address: {}", &bind_address))
     .run()
     .await
 }
+
