+++ src/lib/rest_api/static/index.html
@@ -29,7 +29,7 @@
         <div id="flex_component">
             <div id="grid_component">
                 <div id="mjpeg">
-                    <img id="fit_img" src="http://localhost:42002/live_streaming" width="500" height="500">
+                    <img id="fit_img" src="/live_streaming" width="500" height="500">
                     <!-- <img id="fit_img" src="https://pngimg.com/uploads/google/google_PNG19632.png"> -->
                     <canvas id="fit_canvas"></canvas>
                 </div>
